@@ -5,7 +5,7 @@ version: 0.0.12
 homepage: https://github.com/137Vishal/rwa_deep_ar
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
